split code in ScreenMessage
