+++ app/src/main/java/com/laoschool/screen/ScreenMessage.java
@@ -1,19 +1,13 @@
 package com.laoschool.screen;
 
 
-import android.annotation.SuppressLint;
-import android.app.Activity;
 import android.content.Context;
 import android.os.Bundle;
 import android.os.Handler;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -23,13 +17,12 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
 import android.widget.ProgressBar;
-import android.widget.TextView;
 
 import com.astuetz.PagerSlidingTabStrip;
 import com.laoschool.R;
-import com.laoschool.adapter.ListMessageAdapter;
+import com.laoschool.screen.pager.MessagesPager;
+import com.laoschool.adapter.MessagesPagerAdapter;
 import com.laoschool.entities.Message;
-import com.laoschool.listener.OnLoadMoreListener;
 import com.laoschool.model.AsyncCallback;
 import com.laoschool.model.DataAccessImpl;
 import com.laoschool.model.DataAccessInterface;
@@ -39,9 +32,7 @@ import com.laoschool.view.FragmentLifecycle;
 import com.laoschool.view.ViewpagerDisableSwipeLeft;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -177,7 +168,7 @@ public class ScreenMessage extends Fragment implements FragmentLifecycle {
         final String status = "";
         final String fromID = ((form_id > 0) ? String.valueOf(form_id) : "");
         Log.d(TAG, "_getDataFormServer()):\n" +
-                "getMessages(classID=" + classID + "\n" +
+                "getMessagesFormServer(classID=" + classID + "\n" +
                 ",fromUserID=" + fromUserID + ",fromDate=" + fromDate + "\n" +
                 ",toUserID=" + toUserID + ",toDate=" + toDate + "\n" +
                 ",channel=" + channel + ",status=" + status + "\n" +
@@ -197,7 +188,7 @@ public class ScreenMessage extends Fragment implements FragmentLifecycle {
                         try {
                             int sizeResults = result.size();
                             Log.d(TAG, "_getDataFormServer()):" +
-                                    "getMessages()/onSuccess() Results size=" + sizeResults);
+                                    "getMessagesFormServer()/onSuccess() Results size=" + sizeResults);
                             for (Message message : result) {
                                 dataAccessMessage.addOrUpdateMessage(message);
                             }
@@ -211,7 +202,7 @@ public class ScreenMessage extends Fragment implements FragmentLifecycle {
                     @Override
                     public void onFailure(String message) {
                         Log.e(TAG, "NotificationList:setOnRefreshListener():\n" +
-                                "getMessages(classID=" + classID + "\n" +
+                                "getMessagesFormServer(classID=" + classID + "\n" +
                                 ",fromUserID=" + fromUserID + ",fromDate=" + fromDate + "\n" +
                                 ",toUserID=" + toUserID + ",toDate=" + toDate + "\n" +
                                 ",channel=" + channel + ",status=" + status + "\n" +
@@ -254,9 +245,9 @@ public class ScreenMessage extends Fragment implements FragmentLifecycle {
         List<Message> messagesFormUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_FROM_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 1);
         // Bind the tabs to the ViewPager
         if (inita)
-            messagesPagerAdapter = new MessagesPagerAdapter(fr, messagesForUserInbox, messagesToUserUnread, messagesFormUser);
+            messagesPagerAdapter = new MessagesPagerAdapter(fr, thiz, messagesForUserInbox, messagesToUserUnread, messagesFormUser);
         else {
-            messagesPagerAdapter = new MessagesPagerAdapter(fr, new ArrayList<Message>(), new ArrayList<Message>(), new ArrayList<Message>());
+            messagesPagerAdapter = new MessagesPagerAdapter(fr, thiz, new ArrayList<Message>(), new ArrayList<Message>(), new ArrayList<Message>());
         }
         pager.setAdapter(messagesPagerAdapter);
         tabs.setViewPager(pager);
@@ -276,19 +267,19 @@ public class ScreenMessage extends Fragment implements FragmentLifecycle {
 
             @Override
             public void onPageSelected(int position) {
-                MessageListFragment notifragment = ((MessagesPagerAdapter) (pager.getAdapter())).getFragment(position);
+                MessagesPager notifragment = ((MessagesPagerAdapter) (pager.getAdapter())).getFragment(position);
                 switch (position) {
                     case 0:
                         List<Message> messagesForUserInbox = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 1);
-                        notifragment._setListMessage(messagesForUserInbox, 0, true);
+                        notifragment.setListMessage(messagesForUserInbox, 0, true);
                         break;
                     case 1:
                         List<Message> messagesForUserUnread = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 0);
-                        notifragment._setListMessage(messagesForUserUnread, 1, true);
+                        notifragment.setListMessage(messagesForUserUnread, 1, true);
                         break;
                     case 2:
                         List<Message> messagesFormUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_FROM_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 1);
-                        notifragment._setListMessage(messagesFormUser, 2, true);
+                        notifragment.setListMessage(messagesFormUser, 2, true);
                         break;
                 }
             }
@@ -336,477 +327,16 @@ public class ScreenMessage extends Fragment implements FragmentLifecycle {
         switch (id) {
             case R.id.action_create_message:
                 iScreenMessage._gotoScreenCreateMessage();
-//                Toast.makeText(getActivity(), "create message", Toast.LENGTH_SHORT).show();
                 return true;
         }
         return super.onOptionsItemSelected(item);
     }
 
     @Override
-    public void onAttach(Activity activity) {
-        Log.d(TAG, "onAttach(Activity)");
-        super.onAttach(activity);
-        iScreenMessage = (IScreenMessage) activity;
-        if (LaoSchoolShared.myProfile == null) {
-            iScreenMessage.reLogin();
-        }
-    }
-
-
-    @SuppressLint("ValidFragment")
-    public static class MessageListFragment extends Fragment {
-        private static String ARG_POSITION = "position";
-        private int position;
-        private RecyclerView mRecyclerListMessage;
-        private Context context;
-        SwipeRefreshLayout mSwipeRefreshLayout;
-        private List<Message> messages;
-
-        public MessageListFragment(int position, List<Message> messages) {
-            this.position = position;
-            this.messages = messages;
-        }
-
-        public MessageListFragment() {
-
-        }
-
-        public static MessageListFragment newInstance(int page) {
-            Bundle args = new Bundle();
-            args.putInt(ARG_POSITION, page);
-            MessageListFragment fragment = new MessageListFragment();
-            fragment.setArguments(args);
-            return fragment;
-        }
-
-        @Override
-        public void onCreate(Bundle savedInstanceState) {
-            super.onCreate(savedInstanceState);
-            // position = getArguments().getInt(ARG_POSITION);
-            this.context = getActivity();
-        }
-
-        @Override
-        public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-            View view = inflater.inflate(R.layout.view_message_list, container, false);
-            TextView lbNoMessage = (TextView) view.findViewById(R.id.lbNoMessage);
-            mSwipeRefreshLayout = (SwipeRefreshLayout) view.findViewById(R.id.swipeRefreshLayout);
-            mRecyclerListMessage = (RecyclerView) view.findViewById(R.id.mRecyclerListMessage);
-            LinearLayoutManager linearLayoutManager = new LinearLayoutManager(context);
-            //set adapter
-            mRecyclerListMessage.setLayoutManager(linearLayoutManager);
-            if (messages.size() > 0) {
-                lbNoMessage.setVisibility(View.GONE);
-                mSwipeRefreshLayout.setVisibility(View.VISIBLE);
-                _setListMessage(messages, position, true);
-                _handlerSwipeReload();
-            } else {
-                lbNoMessage.setVisibility(View.VISIBLE);
-                mSwipeRefreshLayout.setVisibility(View.GONE);
-            }
-
-
-
-            return view;
-        }
-
-        private void _handlerSwipeReload() {
-            mSwipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
-                @Override
-                public void onRefresh() {
-                    getMessages(position);
-                    // Refresh items
-                    mSwipeRefreshLayout.setRefreshing(false);
-                }
-            });
-        }
-
-        private void getMessages(final int position) {
-            Log.d(TAG, "getMessages() -position=" + position);
-            int form_id = DataAccessMessage.getMaxMessagesID(LaoSchoolShared.myProfile.getId());
-
-            final String classID = "";
-            final String fromUserID = ((position == 2) ? String.valueOf(LaoSchoolShared.myProfile.getId()) : "");
-            final String fromDate = "";
-            final String toUserID = ((position == 0 || position == 1) ? String.valueOf(LaoSchoolShared.myProfile.getId()) : "");
-            final String toDate = "";
-            final String channel = "";
-            final String status = "";
-            final String fromID = ((form_id > 0) ? String.valueOf(form_id) : "");
-            Log.d(TAG, "getMessages(classID=" + classID + "\n" +
-                    ",fromUserID=" + fromUserID + ",fromDate=" + fromDate + "\n" +
-                    ",toUserID=" + toUserID + ",toDate=" + toDate + "\n" +
-                    ",channel=" + channel + ",status=" + status + "\n" +
-                    ",fromID=" + fromID + ")");
-            service.getMessages(
-                    classID//classID
-                    , fromUserID//from user ID
-                    , fromDate//from date
-                    , toDate//to date
-                    , toUserID//to user ID
-                    , channel//channel
-                    , status//status
-                    , fromID//from id
-                    , new AsyncCallback<List<Message>>() {
-                        @Override
-                        public void onSuccess(final List<Message> result) {
-                            try {
-                                int sizeResults = result.size();
-                                Log.d(TAG, "getMessages()/onSuccess() Results size=" + sizeResults);
-                                for (Message message : result) {
-                                    dataAccessMessage.addOrUpdateMessage(message);
-                                }
-                                swapData();
-                            } catch (Exception e) {
-                                Log.d(TAG, "getMessages() onSuccess Exception=" + e.getMessage());
-                            }
-
-                        }
-
-                        @Override
-                        public void onFailure(String message) {
-                            Log.e(TAG, "getMessages(classID=" + classID + "\n" +
-                                    ",fromUserID=" + fromUserID + ",fromDate=" + fromDate + "\n" +
-                                    ",toUserID=" + toUserID + ",toDate=" + toDate + "\n" +
-                                    ",channel=" + channel + ",status=" + status + "\n" +
-                                    ",fromID=" + fromID + ")/onFailure():" + message);
-                        }
-
-                        @Override
-                        public void onAuthFail(String message) {
-                            LaoSchoolShared.goBackToLoginPage(context);
-                        }
-                    });
-        }
-
-        private void swapData() {
-            List<Message> messagesForUser = new ArrayList<>();
-            if (LaoSchoolShared.myProfile != null) {
-                if (position == 0) {
-                    messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 1);
-                } else if (position == 1) {
-                    messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 0);
-                } else if (position == 2) {
-                    messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_FROM_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 1);
-                }
-                Log.d(TAG, "MessageListFragment:getListMessagesForUser size=" + messagesForUser.size());
-            }
-            _setListMessage(messagesForUser, position, true);
-        }
-
-//        private void _defineListMessage() {
-//            //Load message in local
-//            int countLocal = dataAccessMessage.getMessagesCount();
-//            Log.d(TAG, "NotificationList:count message in Local=" + countLocal);
-////
-//            if (countLocal > 0) {
-//                _getListMessageFormLocalData();
-//            } else {
-//                _getListMessageFormServer();
-//            }
-//        }
-
-//        private void _getListMessageFormServer() {
-//            Log.d(TAG, "NotificationList:_getListMessageFormServer() position=" + position);
-//            int form_id = DataAccessMessage.getMaxMessagesID(LaoSchoolShared.myProfile.getId());
-//
-
-//            final String classID = "";
-//            final String fromUserID = ((position == 2) ? String.valueOf(LaoSchoolShared.myProfile.getId()) : "");
-//            final String fromDate = "";
-//            final String toUserID = ((position == 0 || position == 1) ? String.valueOf(LaoSchoolShared.myProfile.getId()) : "");
-//            final String toDate = "";
-//            final String channel = "";
-//            final String status = "";
-//            final String fromID = ((form_id > 0) ? String.valueOf(form_id) : "");
-//            service.getMessages(
-//                    classID//classID
-//                    , fromUserID//from user ID
-//                    , fromDate//from date
-//                    , toDate//to date
-//                    , toUserID//to user ID
-//                    , channel//channel
-//                    , status//status
-//                    , fromID//from id
-//                    , new AsyncCallback<List<Message>>() {
-//                        @Override
-//                        public void onSuccess(List<Message> result) {
-//                            try {
-//                                int sizeResults = result.size();
-//                                Log.d(TAG, "NotificationList:setOnRefreshListener():\n" +
-//                                        "getMessages(classID=" + classID + "\n" +
-//                                        ",fromUserID=" + fromUserID + ",fromDate=" + fromDate + "\n" +
-//                                        ",toUserID=" + toUserID + ",toDate=" + toDate + "\n" +
-//                                        ",channel=" + channel + ",status=" + status + "\n" +
-//                                        ",fromID=" + fromID + ")/onSuccess() Results size=" + sizeResults);
-//                                for (Message message : result) {
-//                                    dataAccessMessage.addOrUpdateMessage(message);
-//                                }
-//                                _getListMessageFormLocalData();
-//                            } catch (Exception e) {
-//                                e.printStackTrace();
-//                            }
-//
-//                        }
-//
-//                        @Override
-//                        public void onFailure(String message) {
-//                            Log.e(TAG, "NotificationList:setOnRefreshListener():\n" +
-//                                    "getMessages(classID=" + classID + "\n" +
-//                                    ",fromUserID=" + fromUserID + ",fromDate=" + fromDate + "\n" +
-//                                    ",toUserID=" + toUserID + ",toDate=" + toDate + "\n" +
-//                                    ",channel=" + channel + ",status=" + status + "\n" +
-//                                    ",fromID=" + fromID + ")/onFailure():" + message);
-//                        }
-//                    });
-//
-//        }
-//
-//        public void _getListMessageFormLocalData() {
-//            Log.d(TAG, "NotificationList:_getListMessageFormLocalData() position=" + position);
-//            List<Message> messagesForUser = new ArrayList<>();
-//            if (LaoSchoolShared.myProfile != null) {
-//                if (position == 0) {
-//                    messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 1);
-//                } else if (position == 1) {
-//                    messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 0);
-//                } else if (position == 2) {
-//                    messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_FROM_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 1);
-//                }
-//                Log.d(TAG, "MessageListFragment:getListMessagesForUser size=" + messagesForUser.size());
-//            }
-//            _setListMessage(messagesForUser, position);
-//        }
-
-
-        private void _getListMessageFormServer() {
-            Log.d(TAG, "NotificationList:_getListMessageFormServer() position=" + position);
-            int form_id = DataAccessMessage.getMaxMessagesID(LaoSchoolShared.myProfile.getId());
-
-            final String classID = "";
-            final String fromUserID = ((position == 2) ? String.valueOf(LaoSchoolShared.myProfile.getId()) : "");
-            final String fromDate = "";
-            final String toUserID = ((position == 0 || position == 1) ? String.valueOf(LaoSchoolShared.myProfile.getId()) : "");
-            final String toDate = "";
-            final String channel = "";
-            final String status = "";
-            final String fromID = ((form_id > 0) ? String.valueOf(form_id) : "");
-            service.getMessages(
-                    classID//classID
-                    , fromUserID//from user ID
-                    , fromDate//from date
-                    , toDate//to date
-                    , toUserID//to user ID
-                    , channel//channel
-                    , status//status
-                    , fromID//from id
-                    , new AsyncCallback<List<Message>>() {
-                        @Override
-                        public void onSuccess(List<Message> result) {
-                            try {
-                                int sizeResults = result.size();
-                                Log.d(TAG, "NotificationList:setOnRefreshListener():\n" +
-                                        "getMessages(classID=" + classID + "\n" +
-                                        ",fromUserID=" + fromUserID + ",fromDate=" + fromDate + "\n" +
-                                        ",toUserID=" + toUserID + ",toDate=" + toDate + "\n" +
-                                        ",channel=" + channel + ",status=" + status + "\n" +
-                                        ",fromID=" + fromID + ")/onSuccess() Results size=" + sizeResults);
-                                for (Message message : result) {
-                                    dataAccessMessage.addOrUpdateMessage(message);
-                                }
-                                _getListMessageFormLocalData();
-                            } catch (Exception e) {
-                                e.printStackTrace();
-                            }
-
-                        }
-
-                        @Override
-                        public void onFailure(String message) {
-                            Log.e(TAG, "NotificationList:setOnRefreshListener():\n" +
-                                    "getMessages(classID=" + classID + "\n" +
-                                    ",fromUserID=" + fromUserID + ",fromDate=" + fromDate + "\n" +
-                                    ",toUserID=" + toUserID + ",toDate=" + toDate + "\n" +
-                                    ",channel=" + channel + ",status=" + status + "\n" +
-                                    ",fromID=" + fromID + ")/onFailure():" + message);
-                        }
-
-                        @Override
-                        public void onAuthFail(String message) {
-                            LaoSchoolShared.goBackToLoginPage(context);
-                        }
-                    });
-
-        }
-
-        public void _getListMessageFormLocalData() {
-            Log.d(TAG, "NotificationList:_getListMessageFormLocalData() position=" + position);
-            List<Message> messagesForUser = new ArrayList<>();
-            if (LaoSchoolShared.myProfile != null) {
-                if (position == 0) {
-                    messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 1);
-                } else if (position == 1) {
-                    messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 0);
-                } else if (position == 2) {
-                    messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_FROM_USR_ID, LaoSchoolShared.myProfile.getId(), 30, 0, 1);
-                }
-                Log.d(TAG, "MessageListFragment:getListMessagesForUser size=" + messagesForUser.size());
-            }
-            _setListMessage(messagesForUser, position, true);
-        }
-
-
-        private void _setListMessage(final List<Message> messages, final int position, boolean swapData) {
-            try {
-
-                final ListMessageAdapter listMessageAdapter = new ListMessageAdapter(mRecyclerListMessage, thiz, messages, position);
-                listMessageAdapter.setOnLoadMoreListener(new OnLoadMoreListener() {
-                                                             @Override
-                                                             public void onLoadMore() {
-                                                                 int countMessageFormLocal = 0;
-                                                                 if (position == 0) {
-                                                                     countMessageFormLocal = dataAccessMessage.getMessagesCountFormUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId());
-                                                                 } else if (position == 1) {
-                                                                     countMessageFormLocal = dataAccessMessage.getMessagesCountFormUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), 0);
-                                                                 } else if (position == 2) {
-                                                                     countMessageFormLocal = dataAccessMessage.getMessagesCountFormUser(Message.MessageColumns.COLUMN_NAME_FROM_USR_ID, LaoSchoolShared.myProfile.getId());
-                                                                 }
-                                                                 if (messages != null) {
-                                                                     if (messages.size() < countMessageFormLocal) {
-                                                                         Log.d(TAG, "onLoadMore()");
-                                                                         messages.add(null);
-                                                                         listMessageAdapter.notifyItemInserted(messages.size() - 1);
-                                                                         _loadMoreData(messages, listMessageAdapter, position);
-                                                                     } else {
-                                                                         Log.d(TAG, "onLoadMore() No message load !!!");
-                                                                     }
-                                                                 } else {
-                                                                     Log.d(TAG, "onLoadMore() message Null");
-                                                                 }
-
-
-                                                             }
-                                                         }
-
-                );
-                if (!swapData)
-                    mRecyclerListMessage.setAdapter(listMessageAdapter);
-                else
-                    mRecyclerListMessage.swapAdapter(listMessageAdapter, true);
-
-            } catch (Exception e) {
-                Log.e(TAG, "NotificationList:_setListMessage():" + e.getMessage());
-            }
-        }
-
-        private void _loadMoreData(final List<Message> messages, final ListMessageAdapter listMessageAdapter, final int position) {
-            //Load more data for reyclerview
-            new Handler().postDelayed(new Runnable() {
-                @Override
-                public void run() {
-                    Log.e(TAG, "Load More 2");
-
-                    //Remove loading item
-                    messages.remove(messages.size() - 1);
-                    listMessageAdapter.notifyItemRemoved(messages.size());
-
-                    //Load data
-
-                    List<Message> messagesForUser = new ArrayList<>();
-                    if (LaoSchoolShared.myProfile != null) {
-                        switch (position) {
-                            case 0:
-                                messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), messages.size() + 30, messages.size(), 1);
-                                break;
-                            case 1:
-                                messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_TO_USR_ID, LaoSchoolShared.myProfile.getId(), messages.size() + 30, messages.size(), 0);
-                                break;
-                            case 2:
-                                messagesForUser = dataAccessMessage.getListMessagesForUser(Message.MessageColumns.COLUMN_NAME_FROM_USR_ID, LaoSchoolShared.myProfile.getId(), messages.size() + 30, messages.size(), 1);
-                                break;
-                        }
-                        Log.d(TAG, "MessageListFragment:getListMessagesForUser size=" + messages.size());
-                    }
-                    messages.addAll(messagesForUser);
-
-                    listMessageAdapter.notifyDataSetChanged();
-                    listMessageAdapter.setLoaded();
-                }
-            }, 2000);
-        }
-
-    }
-
-    public static class MessagesPagerAdapter extends FragmentPagerAdapter {
-        private FragmentManager mFragmentManager;
-        private Map<Integer, String> mFragmentTags;
-        private List<Message> messagesForUserInbox;
-        private List<Message> messagesToUserUnread;
-        private List<Message> messagesFormUser;
-
-
-        public MessagesPagerAdapter(FragmentManager fr, List<Message> messagesForUserInbox, List<Message> messagesToUserUnread, List<Message> messagesFormUser) {
-            super(fr);
-            mFragmentManager = fr;
-            mFragmentTags = new HashMap<Integer, String>();
-            this.messagesForUserInbox = messagesForUserInbox;
-            this.messagesToUserUnread = messagesToUserUnread;
-            this.messagesFormUser = messagesFormUser;
-        }
-
-        @Override
-        public CharSequence getPageTitle(int position) {
-            if (position == 0)
-                return "Inbox";
-            else if (position == 1)
-                return "Unread";
-            else if (position == 2)
-                return "Send";
-            else
-                return null;
-        }
-
-        @Override
-        public int getCount() {
-            return 3;
-        }
-
-        @Override
-        public Fragment getItem(int position) {
-            if (position == 0) {
-                return new MessageListFragment(0, messagesForUserInbox);
-            } else if (position == 1) {
-                return new MessageListFragment(1, messagesToUserUnread);
-            } else if (position == 2) {
-                return new MessageListFragment(2, messagesFormUser);
-            } else {
-                return null;
-            }
-        }
-
-        public int getItemPosition(Object object) {
-            return POSITION_NONE;
-        }
-
-        @Override
-        public Object instantiateItem(ViewGroup container, int position) {
-            Object obj = super.instantiateItem(container, position);
-            if (obj instanceof Fragment) {
-                // record the fragment tag here.
-                Fragment f = (Fragment) obj;
-                String tag = f.getTag();
-                mFragmentTags.put(position, tag);
-            }
-            return obj;
-        }
-
-        public MessageListFragment getFragment(int position) {
-            String tag = mFragmentTags.get(position);
-            if (tag == null)
-                return null;
-            return (MessageListFragment) mFragmentManager.findFragmentByTag(tag);
-        }
-
+    public void onAttach(Context context) {
+        Log.d(TAG, "onAttach(context)");
+        super.onAttach(context);
+        iScreenMessage = (IScreenMessage) context;
     }
 
     @Override
@@ -817,12 +347,6 @@ public class ScreenMessage extends Fragment implements FragmentLifecycle {
             HomeActivity homeActivity = (HomeActivity) getActivity();
             homeActivity.logoutApplication();
         }
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        Log.d(TAG, "onAttach(Context)");
-        super.onAttach(context);
     }
 
     @Override
@@ -860,10 +384,4 @@ public class ScreenMessage extends Fragment implements FragmentLifecycle {
         Log.d(TAG, "onDetach()");
         super.onDetach();
     }
-
-//    @Override
-//    public void setUserVisibleHint(boolean isVisibleToUser) {
-//        Log.d(TAG, "setUserVisibleHint(" + isVisibleToUser + ")");
-//        super.setUserVisibleHint(isVisibleToUser);
-//    }
 }
